Day 7: Add an UNDEF operator. Don't init with inputs.
There's no guarantee a gate's inputs will be defined when the gate is.
One way to handle this is to use an UNDEF gate for input.  Later it
is replaced with the defined gate.  This allows gates to be stored,
by name, in a hash table, even the UNDEF ones.  As UNDEF gates
are defined they can be found in the hash table and overwritten,
avoiding having to find where they go in the circuit.

inputs aren't put in as part of initialization both to allow gates
without inputs (should I choose another technique for day 7 later).
It also avoids the ambiguity about if an input is a gate or a number.
There are now separate methods for setting inputs and setting values.