Change Graph_shortest_route_cost* to return GraphCost.
This was a UI convenience that lead to a bug.

Infinity cannot be an integer. Once Graph_shortest_route_cost() started
calling Graph_shortest_route_cost_from() if there was no route from a starting
point the infinite cost was silently being cast to -MAX_INT which
was coming through as the cheapest cost.