Simplify Graph_min_cost().
Since we're recursively working backwards from a complete solution,
it's not from start to next, it's from start to the "current" node
we're going to work backwards from.

This means that node has already been visited which makes setting up
and working with visited easier.  Graph_min_cost(), not the caller,
is responsible for removing the current node from visited.

The symmetrical optimization is now to just flip start and current,
both are visited.